Extracted new interface to own file
